Refactor parsing into separate file
